[fix]: refactoring the UserService class, method __construct has 7 arguments (exceeds 4 allowed), remove the DefaultService.php
